divide measures into sources
